Refactor import statements to remove redundant 'wan_animate' prefix
